Fix a typo in README and rename ansiblecontentparser to ansible_content_parser
