+++ src/ansible_content_parser/__init__.py
@@ -1,0 +1,6 @@
+"""Main ansible_content_parser package."""
+from __future__ import annotations
+
+from ansible_content_parser.version import __version__
+
+__all__ = ("__version__",)
