--- src/ansiblecontentparser/__init__.py
@@ -1,6 +1,0 @@
-"""Main ansiblecontentparser package."""
-from __future__ import annotations
-
-from ansiblecontentparser.version import __version__
-
-__all__ = ("__version__",)
